# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/create:
    x-swagger-router-controller: users
    post:
      operationId: createUser
      tags:
        - user
      consumes:
        - application/json
      produces:
        - application/json
      description: Create a new user
      parameters:
        - name: name
          in: body
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/list:
      x-swagger-router-controller: users
      get:
        tags:
          - user
        description: List of users
        operationId: listUsers
        produces:
          - application/json
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/UserResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  HelloResponse:
    required:
      - message
    properties:
      message:
        type: string
  User:
    required:
      - name
    properties:
      name:
        type: string
  UserResponse:
    required:
      - message
      - sucess
    properties:
      message:
        type: string
      sucess:
        type: boolean
      users:
        type: array
        items: {}

  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string